chore(Frontend folder): changes in layout
