+++ frontend/src/components/AppLayout/Buscador.jsx
