+++ frontend/src/components/AppLayout/Navbar.jsx
@@ -1,10 +1,10 @@
 import styles from "./Navbar.module.css";
-import cat from "../assets/Cat.png";
+import cat from "../../assets/Cat.png";
 import LogOut from "./LogOut";
 import { useContext, useEffect, useState } from "react";
-import { context } from "../App";
-import { getUserSession } from "../functions/localStorage";
-import { userApi } from "../functions/apiWrapper";
+import { context } from "../../App";
+import { getUserSession } from "../../functions/localStorage";
+import { userApi } from "../../functions/apiWrapper";
 import { useNavigate } from "react-router-dom";
 import { Home, UserRound, Mail, BellRing, Bookmark } from "lucide-react";
