+++ frontend/src/components/Meows/ListOfMeows.jsx
@@ -1,15 +1,16 @@
-import { useState, useEffect, useContext } from "react";
-import styles from "../pages/Meows.module.css";
-import { meowApi, userApi } from "../functions/apiWrapper";
-import user from "../assets/user.png";
-import LikeButton from "./LikeButton";
-import { getUserSession, getUserToken } from "../functions/localStorage";
-import Loading from "../effects/Loading";
+import { useState, useContext, useEffect } from "react";
+import { meowApi, userApi } from "../../functions/apiWrapper.js";
+import styles from "./ListOfMeows.module.css";
+import user from "../../assets/user.png";
+import Loading from "../../effects/Loading.jsx";
+import { context } from "../../App.jsx";
+import { getUserToken } from "../../functions/localStorage.js";
+import LikeButton from "../Buttons/LikeButton.jsx";
+import RepostMeow from "./RepostMeow.jsx";
 import { useNavigate } from "react-router-dom";
-import { context } from "../App";
-import { formatMeowDate } from "../functions/dateFormat";
+import { formatMeowDate } from "../../functions/dateFormat.js";
 
-const MeowsLiked = () => {
+function Meows() {
   const [meows, setMeows] = useState("");
   const [error, setError] = useState(false);
   const [loading, setLoading] = useState(false);
@@ -22,17 +23,15 @@ const MeowsLiked = () => {
   useEffect(() => {
     const getAllMeows = async () => {
       try {
-        const { id } = getUserSession();
         const token = getUserToken();
         setLoading(true);
-        const res = await meowApi().get(`likes/${id}`, {
+        const res = await meowApi().get("/", {
           headers: {
             Authorization: `Bearer ${token}`,
           },
         });
         setLoading(false);
         const data = res.data;
-
         setMeows(data.reverse());
 
         const uniqueAuthorIds = Array.from(
@@ -79,6 +78,7 @@ const MeowsLiked = () => {
 
         setMeows(meowsToShow);
       } catch (error) {
+        console.error(error);
         setError(true);
         seterrorMessage(error.message);
       }
@@ -91,7 +91,7 @@ const MeowsLiked = () => {
   if (error)
     return (
       <div style={{ fontSize: "40px" }}>
-        Oops, something went wrong!
+        Ops, something went wrong!
         <p style={{ fontSize: "20px", color: "red", fontWeight: "bold" }}>
           {errorMessage}
         </p>
@@ -103,34 +103,52 @@ const MeowsLiked = () => {
       {meows &&
         meows.map((meow) => {
           return (
-            <div className={styles.container}>
-              <div className={styles.userContainer}>
-                <img src={user} />
-                <p
-                  onClick={() => {
-                    navigate("/user/" + meow.authorUsername);
-                    reload.setReload(!reload.reload);
+            <div key={meow._id} className={styles.container}>
+              <div className={styles.meowsContainer}>
+                <div className={styles.userContainer}>
+                  <img src={user} />
+                  {!meow.repostedMeowId && (
+                    <>
+                      <p
+                        onClick={() => {
+                          navigate("/user/" + meow.authorUsername);
+                          reload.setReload(!reload.reload);
+                        }}
+                        className={styles.nameSurname}
+                      >
+                        {meow.nameAuthor} {meow.surnameAuthor}
+                      </p>
+                      <p className={styles.username}>@{meow.authorUsername}</p>
+                    </>
+                  )}
+                  {meow.repostedMeowId && (
+                    <div>
+                      <p
+                        className={styles.repost}
+                        style={{ fontSize: "0.8rem" }}
+                      >
+                        Reposted by: {meow.authorUsername}
+                      </p>
+                      <p>{meow.originalUsername}</p>
+                    </div>
+                  )}
+                </div>
+                <div
+                  onClick={(e) => {
+                    if (e.target.id === "likeButton") return;
+                    navigate(`/meow/${meow._id}`, { state: { meow } });
                   }}
-                  className={styles.nameSurname}
+                  key={meow._id}
+                  className={styles.postContainer}
                 >
-                  {meow.nameAuthor} {meow.surnameAuthor}
-                </p>
-                <p className={styles.username}>@{meow.authorUsername}</p>
-              </div>
-              <div
-                onClick={(e) => {
-                  if (e.target.id === "likeButton") return;
-                  navigate(`/meow/${meow._id}`, { state: { meow } });
-                }}
-                key={meow._id}
-                className={styles.postContainer}
-              >
-                <p>{meow.text}</p>
+                  <p>{meow.text}</p>
+                </div>
                 <div className={styles.likesContainer}>
                   <p>
                     <LikeButton meow={meow} />
+                    <RepostMeow meow={meow} />
+                    <p>{meow.date}</p>
                   </p>
-                  <p>{meow.date}</p>
                 </div>
               </div>
             </div>
@@ -138,6 +156,5 @@ const MeowsLiked = () => {
         })}
     </div>
   );
-};
-
-export default MeowsLiked;
+}
+export default Meows;
