+++ frontend/src/components/Profile/EditProfileForm.jsx
@@ -1,7 +1,7 @@
 import styles from "./EditProfileForm.module.css";
 import { useForm } from "react-hook-form";
 import React, { useState } from "react";
-import { userApi } from "../functions/apiWrapper";
+import { userApi } from "../../functions/apiWrapper";
 import Swal from "sweetalert2";
 
 const EditProfileForm = ({
