+++ frontend/src/components/PublicHome/LoginForm.jsx
@@ -2,10 +2,10 @@ import styles from "./LoginForm.module.css";
 import { useForm } from "react-hook-form";
 import React, { useState } from "react";
 import Swal from "sweetalert2";
-import { userApi } from "../functions/apiWrapper";
-import { setUserSession } from "../functions/localStorage";
-import Loading from "../effects/Loading";
-import { context } from "../App";
+import { userApi } from "../../functions/apiWrapper";
+import { setUserSession } from "../../functions/localStorage";
+import Loading from "../../effects/Loading";
+import { context } from "../../App";
 import { useContext } from "react";
 
 const LoginForm = ({ close, change, load }) => {
