+++ frontend/src/components/PublicHome/PublicHome.jsx
@@ -1,5 +1,5 @@
 import styles from "./PublicHome.module.css";
-import Cat from "../assets/Cat.png";
+import Cat from "../../assets/Cat.png";
 import RegisterForm from "./RegisterForm";
 import LoginForm from "./LoginForm";
 import { useState } from "react";
